Redesigned user location service

+++ Services/Services.API/UserLocation/IUserLocationApi.cs
@@ -6,7 +6,7 @@ namespace Services.API.UserLocation
 {
     public interface IUserLocationApi
     {
-        Task SendCurrentLocationAsync(int routeId, string latitude, string longitude, CancellationToken token);
+        Task SendCurrentLocationAsync(long routeId, string latitude, string longitude, CancellationToken token);
 
         Task<GoogleRouteInfo> GetRouteInfoAsync(string orderId, string userId, string startLatitude, string startLongitude,
             string endLatitude, string endLongitude, CancellationToken token);
