+++ Services/Services.API/UserLocation/UserLocationApi_mock.cs
@@ -8,7 +8,7 @@ namespace Services.API.UserLocation
 {
     public class UserLocationApi_mock : IUserLocationApi
     {
-        public async Task SendCurrentLocationAsync(int routeId, string latitude, string longitude, CancellationToken token)
+        public async Task SendCurrentLocationAsync(long routeId, string latitude, string longitude, CancellationToken token)
         {
             await Task.Delay(1000);
             System.Diagnostics.Debug.WriteLine($"Saved route ({routeId}) coordinate: {latitude}:{longitude}");
